requirements.txt
google-generativeai>=0.3.0
qdrant-client>=1.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.4.2
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
